rc-car/requirements.txt
aiortc>=1.9.0
aiohttp>=3.9.0
python-socketio[asyncio_client]>=5.11.0
numpy>=1.26.0
opencv-python-headless>=4.8.1.78
av